% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{data}
\name{Hts}
\alias{Hts}
\alias{$.Hts}
\title{Hierarchical Time Series}
\format{
An object of class \code{environment} of length 8.
}
\usage{
Hts
}
\description{
Main data structure for working with hierarchical and/or grouped time series.
Holds bottom-level data along with computed hierarchy tree and summation matrix.
}
\section{Methods}{

\subsection{Method \code{new}}{
Create a new HierarchicalTimeSeries
\subsection{Arguments}{
\describe{
\item{\code{bottom_level}}{Dataframe containing the time series data}
\item{\code{spec}}{HierarchySpec object defining the structure}
\item{\code{time_col}}{Name of the time/period column}
\item{\code{value_col}}{Name of the value column}
}}
\subsection{description}{
Instantiate a new HierarchicalTimeSeries
}
\subsection{return}{
A new HierarchicalTimeSeries object
}
}

\subsection{Method \code{from_csv}}{
Load from CSV file
\subsection{Arguments}{
\describe{
\item{\code{path}}{Path to CSV file}
\item{\code{spec}}{HierarchySpec object defining the structure}
\item{\code{time_col}}{Name of the time/period column}
\item{\code{value_col}}{Name of the value column}
}}
\subsection{return}{
A new HierarchicalTimeSeries object
}
}

\subsection{Method \code{n_series}}{
Get total number of series (all aggregation levels)
\subsection{return}{
Integer count of total series
}
}

\subsection{Method \code{n_bottom}}{
Get number of bottom-level series
\subsection{return}{
Integer count of bottom-level series
}
}

\subsection{Method \code{n_periods}}{
Get number of time periods
\subsection{return}{
Integer count of time periods
}
}

\subsection{Method \code{print}}{
Print summary of the hierarchical structure
}

\subsection{Method \code{summation_matrix}}{
Get summation matrix with row and column labels
\subsection{return}{
List containing the matrix and its labels
}
}

\subsection{Method \code{aggregated_series}}{
Get aggregated series
\subsection{return}{
Dataframe containing all series with their hierarchical labels
}
}
}

\examples{
## ---- Method `new` ---- ##
hts_data <- data.frame(
state = c(
rep("Rio de Janeiro", 4), rep("S達o Paulo", 4),
rep("Rio de Janeiro", 4), rep("S達o Paulo", 4)
),
city = c(
rep("Rio de Janeiro", 2), rep("Duque de Caxias", 2),
rep("S達o Paulo", 2), rep("Campinas", 2),
rep("Rio de Janeiro", 2), rep("Duque de Caxias", 2),
rep("S達o Paulo", 2), rep("Campinas", 2)
),
sector = c(
rep("Industry", 8), rep("Agriculture", 8)
),
quarter = c(
rep("2024 Q1", 16), rep("2024 Q2", 16)
),
gdp = c(
1000, 500, 150, 120,
2000, 800, 300, 200,
1500, 800, 200, 150,
2200, 900, 400, 300,
1100, 600, 180, 130,
2100, 850, 320, 220,
1600, 850, 220, 160,
2300, 950, 420, 320
)
)
spec <- HtsSpec$new(c("state", "city"), c("sector"))
hts <- Hts$new(hts_data, spec, "quarter", "gdp")

## ---- Method `from_csv` ---- ##
\dontrun{
write.csv(tsibble::tourism, "data.csv", row.names = FALSE)
spec <- HierarchySpec$new(c("State", "Region"), c("Purpose"))
hts <- HierarchicalTimeSeries$from_csv("data.csv", spec, "Quarter", "Trips")
}


}
\keyword{datasets}
